v1.2: bi-directional list functions

+++ lists/ft_lstsort.c
@@ -6,14 +6,14 @@
 /*   By: samartin <samartin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/27 13:09:00 by samartin          #+#    #+#             */
-/*   Updated: 2023/02/03 14:22:22 by samartin         ###   ########.fr       */
+/*   Updated: 2023/03/17 10:31:21 by samartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "lists.h"
 
 /**
- * It sorts a linked list of integers in a configurable order.
+ * It sorts a linked list of long integers in a configurable order.
  * 
  * @param lst a pointer to the first element of a linked list.
  * @param cmp a function to be added, that takes two ints and returns 0 if they
@@ -21,7 +21,7 @@
  * 
  * @return A pointer to the first element of the sorted list.
  */
-t_list	*ft_lstsort(t_list *lst, int (*cmp)(int, int))
+t_list	*ft_lstsort(t_list *lst, int (*cmp)(long, long))
 {
 	t_list	*cur;
 	t_list	*past_last;
@@ -35,7 +35,7 @@ t_list	*ft_lstsort(t_list *lst, int (*cmp)(int, int))
 		{
 			while (cur->next != past_last)
 			{
-				if (!(cmp(*(int *)cur->content, *(int *)cur->next->content)))
+				if (!(cmp(*(long *)cur->content, *(long *)cur->next->content)))
 				{
 					swp = cur->content;
 					cur->content = cur->next->content;
